refactor: move Promise types into its own folder
