+++ src/index.ts
@@ -12,7 +12,6 @@ export * from './function'
 export * from './getField'
 export * from './Id'
 export * from './isConstructor'
-export * from './isPromise'
 export * from './JSONTypes'
 export * from './KeyofOptional'
 export * from './KeysWithDiffType'
@@ -23,8 +22,7 @@ export * from './Omit'
 export * from './omitFn'
 export * from './Partial'
 export * from './PrimitiveTypes'
-export * from './PromiseValue'
-export * from './PromiseValueMerge'
+export * from './promise'
 export * from './record'
 export * from './RecursiveIntersect'
 export * from './RecursivePartial'
