+++ src/promise/PromiseValue.spec.ts
@@ -1,4 +1,4 @@
-import { assertType, PromiseValue } from '.'
+import { assertType, PromiseValue } from '..'
 
 test('extract value from Promise', () => {
   const y: PromiseValue<Promise<string>> = '' as any
