+++ src/promise/PromiseValueMerge.spec.ts
@@ -1,4 +1,4 @@
-import { assertType, PromiseValueMerge } from '.'
+import { assertType, PromiseValueMerge } from '..'
 
 test('merge promise value', async () => {
   const result = {} as PromiseValueMerge<Promise<{ a: string }>, Promise<{ b: string }>>
