+++ src/promise/index.ts
@@ -1,0 +1,3 @@
+export * from './isPromise'
+export * from './PromiseValue'
+export * from './PromiseValueMerge'
