+++ src/promise/isPromise.ts
